Refactor: extract common code into utils object
